fix: split builder.io init into client and server logic
