+++ .env.example
@@ -151,7 +151,7 @@ THIRDWEB_ENGINE_ACCESS_TOKEN=
 NEXT_PUBLIC_VERCEL_URL=
 
 # BuilderIO
-BUILDER_IO_PUBLIC_KEY='123'
+NEXT_PUBLIC_BUILDER_IO_PUBLIC_KEY='123'
 BUILDER_IO_PRIVATE_KEY='123'
 
 # Github repository read access token
