+++ .github/workflows/cypress.yml
@@ -80,7 +80,7 @@ jobs:
       STAND_WITH_CRYPTO_SUPPORTER_NFT_CONTRACT_ADDRESS: ${{ vars.STAND_WITH_CRYPTO_SUPPORTER_NFT_CONTRACT_ADDRESS }}
       LA_CRYPTO_EVENT_2024_03_04_NFT_CONTRACT_ADDRESS: ${{ vars.LA_CRYPTO_EVENT_2024_03_04_NFT_CONTRACT_ADDRESS }}
       SWC_I_VOTED_NFT_CONTRACT_ADDRESS: ${{ vars.SWC_I_VOTED_NFT_CONTRACT_ADDRESS }}
-      BUILDER_IO_PUBLIC_KEY: ${{vars.BUILDER_IO_PUBLIC_KEY}}
+      NEXT_PUBLIC_BUILDER_IO_PUBLIC_KEY: ${{vars.NEXT_PUBLIC_BUILDER_IO_PUBLIC_KEY}}
       NEXT_PUBLIC_GOOGLE_TAG_MANAGER_ID: ${{vars.NEXT_PUBLIC_GOOGLE_TAG_MANAGER_ID}}
       # org level secrets
       CYPRESS_RECORD_KEY: ${{ secrets.CYPRESS_RECORD_KEY }}
