+++ .github/workflows/pr.yml
@@ -65,7 +65,7 @@ jobs:
       SWC_DOT_ETH_WALLET: ${{vars.SWC_DOT_ETH_WALLET}}
       LEGACY_NFT_DEPLOYER_WALLET: ${{vars.LEGACY_NFT_DEPLOYER_WALLET}}
       I_AM_A_VOTER_NFT_CONTRACT_ADDRESS: ${{ vars.I_AM_A_VOTER_NFT_CONTRACT_ADDRESS }}
-      BUILDER_IO_PUBLIC_KEY: ${{vars.BUILDER_IO_PUBLIC_KEY}}
+      NEXT_PUBLIC_BUILDER_IO_PUBLIC_KEY: ${{vars.NEXT_PUBLIC_BUILDER_IO_PUBLIC_KEY}}
 
     steps:
       - name: Checkout code
