+++ src/utils/server/serverCMS/models/data/events.ts
@@ -1,17 +1,58 @@
 import * as Sentry from '@sentry/nextjs'
 import pRetry from 'p-retry'
 
-import { builderIOClient } from '@/utils/server/builderIO/client'
+import { serverCMS } from '@/utils/server/serverCMS/serverCMS'
 import { SWCEvents, zodEventSchemaValidation } from '@/utils/shared/getSWCEvents'
 import { getLogger } from '@/utils/shared/logger'
 import { NEXT_PUBLIC_ENVIRONMENT } from '@/utils/shared/sharedEnv'
+
+const logger = getLogger(`builderIOEvents`)
+
+export async function getEvent(eventSlug: string, state: string) {
+  try {
+    const entry = await pRetry(
+      () =>
+        serverCMS.get('events', {
+          query: {
+            data: {
+              slug: eventSlug,
+              state: state.toUpperCase(),
+              isOccuring: true,
+            },
+            ...(NEXT_PUBLIC_ENVIRONMENT === 'production' && { published: 'published' }),
+          },
+          includeUnpublished: NEXT_PUBLIC_ENVIRONMENT !== 'production',
+          cacheSeconds: 60,
+        }),
+      {
+        retries: 3,
+        minTimeout: 5000,
+      },
+    )
+
+    const parsedEntry = zodEventSchemaValidation.safeParse(entry)
+
+    if (!parsedEntry.success) {
+      return null
+    }
+
+    return parsedEntry.data as SWCEvents[0]
+  } catch (e) {
+    logger.error('error getting single event:' + e)
+    Sentry.captureException(e, {
+      level: 'error',
+      tags: { domain: 'getEvents' },
+    })
+    return null
+  }
+}
 
 const LIMIT = 100
 
 async function getAllEventsWithOffset(offset: number) {
   return await pRetry(
     () =>
-      builderIOClient.getAll('events', {
+      serverCMS.getAll('events', {
         query: {
           ...(NEXT_PUBLIC_ENVIRONMENT === 'production' && { published: 'published' }),
           data: {
@@ -30,7 +71,6 @@ async function getAllEventsWithOffset(offset: number) {
   )
 }
 
-const logger = getLogger(`builderIOEvents`)
 export async function getEvents() {
   try {
     let offset = 0
@@ -53,7 +93,7 @@ export async function getEvents() {
 
     return filteredIncompleteEvents
   } catch (e) {
-    logger.error('error getting events entries:' + e)
+    logger.error('error getting all events:' + e)
     Sentry.captureException(e, {
       level: 'error',
       tags: { domain: 'getEvents' },
